(refactor): move chat room socket logic to own file.
